requirements.txt
altgraph==0.17.4
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
boto3==1.28.37
botocore==1.31.37
Brotli==1.1.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
comm==0.1.3
contourpy==1.2.1
cookiecutter==1.7.3
crispy-bootstrap4==2022.1
cycler==0.12.1
debugpy==1.6.7
decorator==4.4.2
distlib==0.3.6
Django==4.2.3
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-filter==23.2
django-jet-reboot==1.3.3
django-js-asset==2.1.0
django-shopping-cart==0.1
django-storages==1.13.2
django-tailwind==2.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
executing==1.2.0
ffmpeg==1.4
filelock==3.12.2
fonttools==4.51.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.5
lml==0.1.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
moviepy==1.0.3
mutagen==1.47.0
nest-asyncio==1.5.6
numpy==1.24.3
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.2.1
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.7.4
platformdirs==3.5.3
poyo==0.5.0
proglog==0.1.10
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.7
psycopg2-binary==2.9.7
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.19.0
pyexcel==0.7.0
pyexcel-io==0.6.6
Pygments==2.15.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.8.0
pyparsing==3.1.2
PyPDF2==3.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.0
requests==2.31.0
s3transfer==0.6.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
texttable==1.7.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
websockets==12.0
whitenoise==6.5.0
wsproto==1.2.0
yt-dlp==2023.11.16
